LayoutTests/platform/glib/imported/w3c/web-platform-tests/fullscreen/api/document-fullscreen-enabled-removing-allowfullscreen.sub-expected.txt

CONSOLE MESSAGE: Feature policy 'Fullscreen' check failed for iframe with origin 'http://www1.web-platform.test:8800' and allow attribute ''.

FAIL Document#fullscreenEnabled removing allowfullscreen after load and then navigating assert_true: document.fullscreenEnabled in the iframe, before navigation expected true got false
